requirements.txt
aws-cdk.aws-s3
aws-cdk.aws-iam
aws-cdk.aws-cloudfront
aws_cdk.aws_certificatemanager
aws_cdk.aws_route53
aws_cdk.aws_route53_targets
